Fixed trash Generation/RecycleLogicIn
I moved the trash data into player manager and the item script now contains the sprites and can now load them and all data has been updated to reflect new sprites.

The trash is stored in a current day dictionary that keeps track of the materials and now can be recycled and it updates of the recycle UI scene.

I also put a trash spawner in the actual game not just my experiment scene.